requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
umap-learn>=0.5.0
xgboost>=1.5.0
hdbscan>=0.8.0
pyyaml>=5.4.0
jinja2>=3.0.0
tqdm>=4.62.0
doubleml>=0.5.0
sentence-transformers>=2.0.0
